This should get you started.  The x matrix is singular when I add GLENN.SEASON which is an ordered Factor
Kosta, here are two models one with the spline (this needs more data prep as it throws a warning on length) and the other is based on categorical data.  if you hist(FNMA$OPTION_LAG1) you will quickly see why I choose the factors.  Also, you will need to set options(contrasts =) and order the factors.  This will allow R to one hot encode you model.  As you can see from summary both the zero incentive and quarter2 are missing.  I suggest more preliminary data analysis and cleaning before you jump directly to the model.  Again, look at the paper (I sent the link) to see how to present data and model.  I understand you are not valuing MBS but the model presentation is the model presentation and it does not change whether one is building the model for your purpose or for valuation.  Indeed, the purpose of the model is simply to generate prepayment vectors and by extension cash flows what and how one works the cash flow is irrelevant - this should get you started on the right path